feat(18228): rename PaymentPlan to PaymentPlanCard
